Plot set of rois using a multicolor scheme by default
